pyproject.toml
[tool.poetry]
name = "storylab-backend"
version = "0.1.0"
description = "Backend API para Storylab con FastAPI, SQLAlchemy y AI Engines"
authors = ["Ruben Serra <tu_email@example.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pydantic = { version = "^2.9.2", extras = ["email"] }
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
psycopg = { version = "^3.2.9", extras = ["binary"] }
python-dotenv = "^1.0.1"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-jose = "^3.3.0"
cryptography = "^43.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
